Refactor language area of settings into its own component
