+++ src/Patient/Settings/index.js
@@ -5,18 +5,15 @@ import OverTopBar from '../Navigation/OverTopBar';
 import { useTranslation } from 'react-i18next';
 import Typography from '@material-ui/core/Typography';
 import Styles from '../../Basics/Styles';
-import Colors from '../../Basics/Colors';
-import ButtonGroup from "@material-ui/core/ButtonGroup";
-import Button from "@material-ui/core/Button";
 import { observer } from 'mobx-react';
 import NewButton from '../../Basics/NewButton';
 import ExitToApp from '@material-ui/icons/ExitToApp';
-import Globe from '@material-ui/icons/Language';
 import PasswordUpdate from '../../Components/PasswordUpdate';
 import PersonalInformation from './PersonalInformation';
 import useLogout from '../../Basics/Logout';
-import { DateTime } from 'luxon';
-import Debugging from './Debugging'
+import Debugging from './Debugging';
+import Language from './Language';
+import Colors from '../../Basics/Colors';
 
 const HealthProfile = observer(() => {
 
@@ -45,10 +42,6 @@ const MainSettings = observer(() => {
     const { t } = useTranslation('translation');
     const logout = useLogout();
 
-    const handleLogout = () => {
-        logout();
-    }
-
     return (
         <>
             <OverTopBar title={t("patient.profile.title")} handleBack={patientUIStore.closeSettings} ></OverTopBar>
@@ -58,35 +51,15 @@ const MainSettings = observer(() => {
                 </div>
                 <Typography className={classes.name} className={classes.name} variant="h2">{patientStore.givenName} {patientStore.familyName}</Typography>
             </div>
-            <LanguageQuestion />
+            <Language />
             <PersonalInformation />
             <Debugging />
             <div className={classes.logoutContainer}>
-                <NewButton onClick={handleLogout} className={classes.logout} icon={<ExitToApp />} text={t("patient.profile.logout")} />
+                <NewButton onClick={logout} className={classes.logout} icon={<ExitToApp />} text={t("patient.profile.logout")} />
             </div>
         </>
     )
 })
-
-const LanguageQuestion = observer(() => {
-    const classes = useStyles();
-    const { uiStore } = useStores();
-    const { t } = useTranslation('translation');
-
-    return (
-        <div className={classes.languageContainer}>
-            <div className={classes.language}>
-                <Globe />
-                <Typography variant="h2">{t("patient.profile.options.language")}</Typography>
-            </div>
-            <ButtonGroup className={classes.group} fullWidth color="primary">
-                <Button onClick={() => { uiStore.setLocale("en") }} className={uiStore.locale === "en" ? classes.selected : classes.default}>{t("patient.profile.options.english")}</Button>
-                <Button onClick={() => { uiStore.setLocale("es-AR") }} className={uiStore.locale === "es-AR" ? classes.selected : classes.default}>{t("patient.profile.options.spanish")}</Button>
-            </ButtonGroup>
-        </div>
-    );
-})
-
 
 const useStyles = makeStyles({
     logout: {
@@ -163,46 +136,6 @@ const useStyles = makeStyles({
         bottom: "0px",
         padding: "5px",
         backgroundColor: "white"
-    },
-    selected: {
-        backgroundColor: Colors.buttonBlue,
-        color: "white",
-        "&:hover": {
-            color: Colors.white,
-            backgroundColor: Colors.accentBlue
-        }
-    },
-    default: {
-        backgroundColor: "white",
-        color: Colors.buttonBlue,
-        "&:hover": {
-            color: Colors.buttonBlue,
-            backgroundColor: Colors.accentBlue
-        }
-    },
-    group: {
-        width: "70%",
-        margin: "1em"
-    },
-    language: {
-        width: "90%",
-        marginLeft: "1em",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "flex-start",
-        "& > svg": {
-            fontSize: "1em",
-            marginRight: "5px"
-        },
-        "& > h2": {
-            fontSize: "1.25em",
-        }
-    },
-    languageContainer: {
-        width: "100%",
-        display: "flex",
-        flexDirection: "column",
-        alignItems: "center"
     },
     pwContainer: {
         width: "90%",
